refactor(react): update number input tests
